requirements.txt
cffi==1.14.5
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.22.1
pycparser==2.20
pycrypto==2.6.1
pymongo==3.11.3
six==1.15.0
Werkzeug==1.0.1
